Extract initialization logic from jemalloc.c into jemalloc_init module
