+++ msvc/projects/vc2015/jemalloc/jemalloc.vcxproj.filters
@@ -208,5 +208,8 @@
     <ClCompile Include="..\..\..\..\src\arenas_management.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\..\..\src\jemalloc_init.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
